⚡ Bolt: Optimize DSP allocation and hot loops
💡 What:
- Pre-allocated `iAcc` and `qAcc` synthesis buffers in `Simulator.swift` to avoid allocating massive dynamic arrays per 0.1s time step.
- Changed `LUT.iq_lut` from `[Int16]` to `[Int]`, bypassing per-sample `Int(...)` casting in the main synthesis loop.
- Optimized carrier phase index bitwise logic (`(carrPhase >> 15) & 0x3fe`) in `generateSamples` to save an unnecessary bit shift.
- Passed pre-allocated buffers as `inout` properties to `GPSSignal.generateSamples`, using fast memory resets (`update(repeating: 0)`).

🎯 Why:
The original approach instantiated multiple heavy arrays (like `iAcc` and `qAcc` buffers scaling to hundreds of megabytes dynamically) on every simulation step, causing excessive memory turnover/ARC overhead. Moreover, the inner loop was constantly casting `Int16` to `Int` and performing redundant bit shifts, wasting millions of CPU cycles.

📊 Impact:
Massively reduces memory allocation pressure across a simulation run (avoiding GBs of GC churn over minutes). Saves native DSP CPU cycles by eliminating unnecessary bit shifts and redundant type casts inside a loop processing millions of samples.

🔬 Measurement:
Run the application over a 5-minute trace and observe memory profiling tools / execution time; CPU wait-time on allocations will vanish.

+++ Sources/gnss-sig-gen-swift/LUT.swift
@@ -19,8 +19,8 @@ import Foundation
 
 /// Look-up tables (LUTs) for signal processing and antenna characteristics.
 struct LUT {
-    /// Interleaved I/Q carrier look-up table (sine/cosine values scaled to Int16).
-    static let iq_lut: [Int16] = [
+    /// Interleaved I/Q carrier look-up table (sine/cosine values scaled to Int).
+    static let iq_lut: [Int] = [
         250,   2, 250,   5, 250,   8, 250,  11, 250,  14, 249,  17, 249,  20, 249,  23,
         249,  26, 248,  29, 248,  32, 248,  35, 247,  38, 247,  41, 246,  44, 245,  47,
         245,  50, 244,  53, 244,  56, 243,  59, 242,  62, 241,  65, 241,  68, 240,  71,
